+++ CHANGES.rst
@@ -7,6 +7,7 @@ mio 0.1.4.dev
 
 - Added "Functions" to the tutorial.
 - Added "Objects" to the tutorial.
+- Added "Traits" to the tutorial.
 
 
 mio 0.1.3 (2013-11-19)